feat: remove persona dashboard pages in favor of customizable dashboard (#292)
Remove ExecutiveView, SecurityView, CiCdView persona tabs from the main
dashboard and the standalone CustomDashboard page. The main dashboard now
serves as the single entry point with:

- Widget-based customizable layout backed by the dashboard config API
- Persona selector to apply recommended widget presets
- Widget catalog to add/remove individual widgets
- /dashboard/custom redirects to /dashboard?view=widgets
