--- frontend/src/pages/CustomDashboard/CustomDashboard.module.css
@@ -1,129 +1,0 @@
-.page {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-}
-
-.toolbar {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-}
-
-.saving {
-  color: var(--fg-muted);
-  font-size: 13px;
-  animation: pulse 1.2s ease-in-out infinite;
-}
-
-@keyframes pulse {
-  0%,
-  100% {
-    opacity: 1;
-  }
-  50% {
-    opacity: 0.5;
-  }
-}
-
-.grid {
-  min-height: 200px;
-}
-
-.widgetWrapper {
-  display: flex;
-  flex-direction: column;
-  border: 1px solid var(--border);
-  border-radius: 10px;
-  background: var(--canvas-default, #fff);
-  overflow: hidden;
-}
-
-.widgetHeader {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  padding: 8px 12px;
-  border-bottom: 1px solid var(--border);
-  background: var(--canvas-subtle);
-}
-
-.dragHandle {
-  cursor: grab;
-  user-select: none;
-  color: var(--fg-muted);
-  font-size: 14px;
-  line-height: 1;
-  padding: 2px 4px;
-  border-radius: 4px;
-}
-
-.dragHandle:active {
-  cursor: grabbing;
-}
-
-.dragHandle:hover {
-  background: rgba(var(--neutral-rgb), 0.12);
-  color: var(--fg);
-}
-
-.widgetTitle {
-  flex: 1;
-  font-size: 13px;
-  font-weight: 600;
-  min-width: 0;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.removeBtn {
-  border: 1px solid var(--border);
-  background: var(--canvas-subtle);
-  color: var(--fg-muted);
-  border-radius: 6px;
-  width: 26px;
-  height: 26px;
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-  font: inherit;
-  font-size: 16px;
-  line-height: 1;
-}
-
-.removeBtn:hover {
-  border-color: var(--danger);
-  color: var(--danger);
-}
-
-.widgetBody {
-  flex: 1;
-  padding: 12px;
-  overflow: auto;
-}
-
-.emptyState {
-  padding: 48px 32px;
-  border: 1px dashed var(--border);
-  border-radius: 12px;
-  text-align: center;
-  color: var(--fg-muted);
-}
-
-.emptyState h3 {
-  margin: 0 0 8px;
-  color: var(--fg);
-  font-size: 18px;
-}
-
-.emptyState p {
-  margin: 0 0 20px;
-}
-
-.emptyActions {
-  display: flex;
-  justify-content: center;
-  gap: 10px;
-}
