--- frontend/src/pages/Dashboard/ExecutiveView.module.css
@@ -1,397 +1,0 @@
-.page {
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-}
-
-.header {
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-start;
-  flex-wrap: wrap;
-  gap: 12px;
-}
-
-.title {
-  font-size: 20px;
-  font-weight: 600;
-}
-
-.subtitle {
-  font-size: 13px;
-  color: var(--fg-muted);
-  margin-top: 2px;
-}
-
-.headerActions {
-  display: flex;
-  align-items: center;
-  gap: 12px;
-}
-
-.periodToggle {
-  display: flex;
-  border: 1px solid var(--border);
-  border-radius: 6px;
-  overflow: hidden;
-}
-
-.periodBtn {
-  padding: 4px 12px;
-  font-size: 12px;
-  border: none;
-  background: var(--canvas-subtle);
-  color: var(--fg-muted);
-  cursor: pointer;
-  font-family: inherit;
-  transition:
-    background 0.15s,
-    color 0.15s;
-}
-
-.periodBtn:not(:last-child) {
-  border-right: 1px solid var(--border);
-}
-
-.periodBtn.active {
-  background: var(--accent);
-  color: var(--fg-on-emphasis);
-}
-
-.periodBtn:hover:not(.active) {
-  background: rgba(var(--neutral-rgb), 0.12);
-}
-
-/* Score card */
-.scoreCard {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  background: var(--canvas-subtle);
-  border: 1px solid var(--border);
-  border-radius: 12px;
-  padding: 20px 24px;
-}
-
-.scoreValue {
-  font-size: 48px;
-  font-weight: 700;
-  color: var(--fg);
-  font-variant-numeric: tabular-nums;
-}
-
-.scoreLabel {
-  font-size: 13px;
-  color: var(--fg-muted);
-  margin-top: 4px;
-}
-
-.scoreDelta {
-  font-size: 16px;
-  font-weight: 600;
-  margin-top: 8px;
-}
-
-.deltaUp {
-  color: var(--success);
-}
-
-.deltaDown {
-  color: var(--danger);
-}
-
-.deltaBigDrop {
-  color: var(--danger);
-  animation: pulse 1.5s ease-in-out infinite;
-}
-
-@keyframes pulse {
-  0%,
-  100% {
-    opacity: 1;
-  }
-  50% {
-    opacity: 0.6;
-  }
-}
-
-/* 3-column grid (score + trend + mom) */
-.grid3 {
-  display: grid;
-  grid-template-columns: minmax(160px, 220px) 1fr 1fr;
-  gap: 16px;
-  align-items: stretch;
-}
-
-/* 2-column grid */
-.grid2 {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  gap: 16px;
-}
-
-/* Compliance */
-.complianceGrid {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(180px, 1fr));
-  gap: 12px;
-}
-
-.complianceCard {
-  padding: 12px;
-}
-
-.complianceFramework {
-  font-size: 14px;
-  font-weight: 600;
-  color: var(--fg);
-  margin-bottom: 8px;
-}
-
-.complianceBar {
-  height: 6px;
-  background: var(--border-muted);
-  border-radius: 3px;
-  overflow: hidden;
-  margin-bottom: 6px;
-}
-
-.complianceFill {
-  height: 100%;
-  background: var(--success);
-  border-radius: 3px;
-  transition: width 0.5s ease;
-}
-
-.complianceMeta {
-  display: flex;
-  justify-content: space-between;
-  font-size: 12px;
-}
-
-.compliancePct {
-  font-weight: 600;
-  color: var(--fg);
-}
-
-.complianceDetail {
-  color: var(--fg-muted);
-}
-
-/* Top risks */
-.riskList {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.riskRow {
-  display: flex;
-  align-items: flex-start;
-  gap: 10px;
-}
-
-.riskDot {
-  width: 10px;
-  height: 10px;
-  border-radius: 50%;
-  flex-shrink: 0;
-  margin-top: 4px;
-}
-
-.riskDot.critical {
-  background: var(--danger);
-}
-
-.riskDot.high {
-  background: var(--severe);
-}
-
-.riskDot.medium {
-  background: var(--attention);
-}
-
-.riskDot.low {
-  background: var(--success);
-}
-
-.riskInfo {
-  flex: 1;
-  min-width: 0;
-}
-
-.riskTitle {
-  font-size: 13px;
-  font-weight: 500;
-  color: var(--fg);
-}
-
-.riskMeta {
-  font-size: 12px;
-  color: var(--fg-muted);
-  margin-top: 2px;
-}
-
-/* Month-over-month */
-.momGrid {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  gap: 12px;
-  padding: 8px 0;
-}
-
-.momCard {
-  text-align: center;
-  padding: 16px;
-  background: var(--canvas-subtle);
-  border-radius: 8px;
-  border: 1px solid var(--border-muted);
-}
-
-.momValue {
-  font-size: 28px;
-  font-weight: 700;
-  color: var(--fg);
-  font-variant-numeric: tabular-nums;
-}
-
-.momLabel {
-  font-size: 12px;
-  color: var(--fg-muted);
-  margin-top: 2px;
-}
-
-.momChange {
-  font-size: 12px;
-  font-weight: 600;
-  margin-top: 6px;
-}
-
-.momUp {
-  color: var(--danger);
-}
-
-.momDn {
-  color: var(--success);
-}
-
-/* Empty state */
-.emptyText {
-  font-size: 13px;
-  color: var(--fg-muted);
-  padding: 16px 0;
-}
-
-.center {
-  display: flex;
-  justify-content: center;
-  padding: 48px;
-}
-
-/* Responsive */
-@media (max-width: 768px) {
-  .grid3 {
-    grid-template-columns: 1fr;
-  }
-
-  .grid2 {
-    grid-template-columns: 1fr;
-  }
-
-  .complianceGrid {
-    grid-template-columns: 1fr;
-  }
-
-  .momGrid {
-    grid-template-columns: 1fr;
-  }
-
-  .scoreCard {
-    width: 100%;
-  }
-
-  .header {
-    flex-direction: column;
-  }
-
-  .headerActions {
-    width: 100%;
-    justify-content: space-between;
-  }
-}
-
-@media (max-width: 480px) {
-  .scoreValue {
-    font-size: 36px;
-  }
-
-  .title {
-    font-size: 18px;
-  }
-}
-
-@media (max-width: 1024px) {
-  .grid3 {
-    grid-template-columns: repeat(2, minmax(0, 1fr));
-  }
-
-  .scoreCard {
-    grid-column: 1 / -1;
-  }
-}
-
-@media (max-width: 768px) {
-  .periodToggle {
-    width: 100%;
-  }
-
-  .periodBtn {
-    flex: 1 1 0;
-  }
-}
-
-@media (prefers-reduced-motion: reduce) {
-  .page,
-  .header,
-  .title,
-  .subtitle,
-  .headerActions,
-  .periodToggle,
-  .periodBtn,
-  .scoreCard,
-  .scoreValue,
-  .scoreLabel,
-  .scoreDelta,
-  .deltaUp,
-  .deltaDown,
-  .deltaBigDrop,
-  .grid3,
-  .grid2,
-  .complianceGrid,
-  .complianceCard,
-  .complianceFramework,
-  .complianceBar,
-  .complianceFill,
-  .complianceMeta,
-  .compliancePct,
-  .complianceDetail,
-  .riskList,
-  .riskRow,
-  .riskDot,
-  .riskInfo,
-  .riskTitle,
-  .riskMeta,
-  .momGrid,
-  .momCard,
-  .momValue,
-  .momLabel,
-  .momChange,
-  .momUp,
-  .momDn,
-  .emptyText,
-  .center {
-    animation: none !important;
-    transition: none !important;
-  }
-}
